Fixed minor errors in testing infrastructure.

+++ encode/tests/mylibs/encode.h
@@ -5,8 +5,8 @@
   #define AN_ENCODE_VALUE(x) __builtin_an_encode_value_i64((x))
   #define AN_DECODE_VALUE(x) __builtin_an_decode_value_i64((x))
 #else
-  #define AN_ENCODE_VALUE(x) (x)
-  #define AN_DECODE_VALUE(x) (x)
+  #define AN_ENCODE_VALUE(x) ((long long)x)
+  #define AN_DECODE_VALUE(x) ((long long)x)
 #endif
 
 #endif /* __ENCODE_H__ */
